extract ConfigurationsManager
